Moved memories to unique component
