+++ client/src/App.css
@@ -40,18 +40,3 @@
   }
 }
 
-.memories-grid {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-}
-
-.memory-form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.memory {
-  border: 1px solid black;
-}
-
