+++ client/src/components/Gallery.js
@@ -1,0 +1,8 @@
+import '../styles/gallery.css'
+export default function Gallery() {
+    return (
+        <div className='Gallery'>
+            <h1>Image Gallery</h1>
+        </div>
+    )
+}
