+++ client/src/components/Landing.js
@@ -1,0 +1,8 @@
+import '../styles/landing.css'
+export default function Landing() {
+    return (
+        <div className='Landing'>
+            <h1>Landing Page</h1>
+        </div>
+    )
+}
