+++ client/src/styles/navbar.css
@@ -1,0 +1,35 @@
+.Navbar {
+    position: fixed;
+    z-index: 1;
+    top: 0;
+    left: 0;
+    background-color: #054b93;
+    color: white;
+    font-family: 'Noto Sans', sans-serif;
+}
+.nav-links {
+    list-style-type: none;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 0 10px 0 10px;
+    transition: all .2s ease-in-out;
+}
+.nav-links li {
+    border: 1px solid white;
+    margin-bottom: 10px;
+}
+
+a {
+    color: white;
+    font-size: 18px;
+    text-decoration: none;
+}
+
+a:focus {
+    text-decoration: underline;
+}
+
+.nav-links li:hover {
+    transform: scale(1.1, 1.1);
+}
